fix(next): fix JsonDocs dts
