+++ package.json
@@ -12,7 +12,6 @@
     "cli/",
     "compiler/",
     "dev-server/",
-    "dist/",
     "internal/",
     "mock-doc/",
     "screenshot/",
